Clean up property utils.
