+++ packages/core/src/utils/index.ts
@@ -4,6 +4,7 @@ export * from './color-utils';
 export * from './file-utils';
 export * from './image-utils';
 export * from './is-plain-object';
-export * from './math-utils';
 export * from './logger';
+export * from './math-utils';
+export * from './property-utils';
 export * from './uuid';
